refactor gerrit client
