+++ prow/BUILD.bazel
@@ -98,7 +98,9 @@ filegroup(
         "//prow/flagutil:all-srcs",
         "//prow/gcsupload:all-srcs",
         "//prow/genfiles:all-srcs",
-        "//prow/gerrit:all-srcs",
+        "//prow/gerrit/adapter:all-srcs",
+        "//prow/gerrit/client:all-srcs",
+        "//prow/gerrit/reporter:all-srcs",
         "//prow/git:all-srcs",
         "//prow/github:all-srcs",
         "//prow/githuboauth:all-srcs",
