+++ prow/cmd/gerrit/BUILD.bazel
@@ -41,7 +41,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//prow/config:go_default_library",
-        "//prow/gerrit:go_default_library",
+        "//prow/gerrit/adapter:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/logrusutil:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
