+++ prow/gerrit/client/BUILD.bazel
@@ -3,25 +3,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["gerrit.go"],
-    importpath = "k8s.io/test-infra/prow/gerrit",
+    importpath = "k8s.io/test-infra/prow/gerrit/client",
     visibility = ["//visibility:public"],
     deps = [
-        "//prow/config:go_default_library",
-        "//prow/kube:go_default_library",
-        "//prow/pjutil:go_default_library",
         "//vendor/github.com/andygrunwald/go-gerrit:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["gerrit_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//prow/config:go_default_library",
-        "//prow/kube:go_default_library",
-        "//vendor/github.com/andygrunwald/go-gerrit:go_default_library",
     ],
 )
 
@@ -34,10 +20,14 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//prow/gerrit/reporter:all-srcs",
-    ],
+    srcs = [":package-srcs"],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["gerrit_test.go"],
+    embed = [":go_default_library"],
+    deps = ["//vendor/github.com/andygrunwald/go-gerrit:go_default_library"],
 )
