+++ prow/gerrit/client/gerrit_test.go
@@ -20,46 +20,24 @@ import (
 	"reflect"
 	"sort"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
-	"github.com/andygrunwald/go-gerrit"
-
-	"k8s.io/test-infra/prow/config"
-	"k8s.io/test-infra/prow/kube"
+	gerrit "github.com/andygrunwald/go-gerrit"
 )
 
-type fca struct {
-	sync.Mutex
-	c *config.Config
-}
-
-func (f *fca) Config() *config.Config {
-	f.Lock()
-	defer f.Unlock()
-	return f.c
-}
-
-type fkc struct {
-	sync.Mutex
-	prowjobs []kube.ProwJob
-}
-
-func (f *fkc) CreateProwJob(pj kube.ProwJob) (kube.ProwJob, error) {
-	f.Lock()
-	defer f.Unlock()
-	f.prowjobs = append(f.prowjobs, pj)
-	return pj, nil
-}
-
 type fgc struct {
-	changes  map[string][]gerrit.ChangeInfo
 	instance string
+	changes  map[string][]gerrit.ChangeInfo
 }
 
 func (f *fgc) QueryChanges(opt *gerrit.QueryChangeOptions) (*[]gerrit.ChangeInfo, *gerrit.Response, error) {
 	changes := []gerrit.ChangeInfo{}
+
+	changeInfos, ok := f.changes[f.instance]
+	if !ok {
+		return &changes, nil, nil
+	}
 
 	project := ""
 	for _, query := range opt.Query {
@@ -70,11 +48,9 @@ func (f *fgc) QueryChanges(opt *gerrit.QueryChangeOptions) (*[]gerrit.ChangeInfo
 		}
 	}
 
-	if changeInfos, ok := f.changes[project]; !ok {
-		return &changes, nil, nil
-	} else {
-		for idx, change := range changeInfos {
-			if idx >= opt.Start && len(changes) <= opt.Limit {
+	for idx, change := range changeInfos {
+		if idx >= opt.Start && len(changes) <= opt.Limit {
+			if project == change.Project {
 				changes = append(changes, change)
 			}
 		}
@@ -93,24 +69,22 @@ func TestQueryChange(t *testing.T) {
 
 	var testcases = []struct {
 		name       string
-		limit      int
 		lastUpdate time.Time
 		changes    map[string][]gerrit.ChangeInfo
-		revisions  []string
+		revisions  map[string][]string
 	}{
 		{
 			name:       "no changes",
-			limit:      2,
 			lastUpdate: now,
-			revisions:  []string{},
+			revisions:  map[string][]string{},
 		},
 		{
 			name:       "one outdated change",
-			limit:      2,
 			lastUpdate: now,
 			changes: map[string][]gerrit.ChangeInfo{
 				"foo": {
 					{
+						Project:         "bar",
 						ID:              "1",
 						CurrentRevision: "1-1",
 						Updated:         now.Add(-time.Hour).Format(layout),
@@ -122,15 +96,15 @@ func TestQueryChange(t *testing.T) {
 					},
 				},
 			},
-			revisions: []string{},
+			revisions: map[string][]string{},
 		},
 		{
 			name:       "one up-to-date change",
-			limit:      2,
 			lastUpdate: now.Add(-time.Minute),
 			changes: map[string][]gerrit.ChangeInfo{
 				"foo": {
 					{
+						Project:         "bar",
 						ID:              "1",
 						CurrentRevision: "1-1",
 						Updated:         now.Format(layout),
@@ -142,15 +116,17 @@ func TestQueryChange(t *testing.T) {
 					},
 				},
 			},
-			revisions: []string{"1-1"},
+			revisions: map[string][]string{
+				"foo": {"1-1"},
+			},
 		},
 		{
 			name:       "one up-to-date change but stale commit",
-			limit:      2,
 			lastUpdate: now.Add(-time.Minute),
 			changes: map[string][]gerrit.ChangeInfo{
 				"foo": {
 					{
+						Project:         "bar",
 						ID:              "1",
 						CurrentRevision: "1-1",
 						Updated:         now.Format(layout),
@@ -162,15 +138,15 @@ func TestQueryChange(t *testing.T) {
 					},
 				},
 			},
-			revisions: []string{},
+			revisions: map[string][]string{},
 		},
 		{
-			name:       "one up-to-date change, wrong project",
-			limit:      2,
+			name:       "one up-to-date change, wrong instance",
 			lastUpdate: now.Add(-time.Minute),
 			changes: map[string][]gerrit.ChangeInfo{
 				"evil": {
 					{
+						Project:         "bar",
 						ID:              "1",
 						CurrentRevision: "1-1",
 						Updated:         now.Format(layout),
@@ -182,15 +158,15 @@ func TestQueryChange(t *testing.T) {
 					},
 				},
 			},
-			revisions: []string{},
+			revisions: map[string][]string{},
 		},
 		{
-			name:       "two up-to-date changes, two projects",
-			limit:      2,
+			name:       "one up-to-date change, wrong project",
 			lastUpdate: now.Add(-time.Minute),
 			changes: map[string][]gerrit.ChangeInfo{
 				"foo": {
 					{
+						Project:         "evil",
 						ID:              "1",
 						CurrentRevision: "1-1",
 						Updated:         now.Format(layout),
@@ -201,8 +177,27 @@ func TestQueryChange(t *testing.T) {
 						},
 					},
 				},
-				"bar": {
+			},
+			revisions: map[string][]string{},
+		},
+		{
+			name:       "two up-to-date changes, two projects",
+			lastUpdate: now.Add(-time.Minute),
+			changes: map[string][]gerrit.ChangeInfo{
+				"foo": {
 					{
+						Project:         "bar",
+						ID:              "1",
+						CurrentRevision: "1-1",
+						Updated:         now.Format(layout),
+						Revisions: map[string]gerrit.RevisionInfo{
+							"1-1": {
+								Created: now.Format(layout),
+							},
+						},
+					},
+					{
+						Project:         "bar",
 						ID:              "2",
 						CurrentRevision: "2-1",
 						Updated:         now.Format(layout),
@@ -214,15 +209,17 @@ func TestQueryChange(t *testing.T) {
 					},
 				},
 			},
-			revisions: []string{"1-1", "2-1"},
+			revisions: map[string][]string{
+				"foo": {"1-1", "2-1"},
+			},
 		},
 		{
 			name:       "one good one bad",
-			limit:      2,
 			lastUpdate: now.Add(-time.Minute),
 			changes: map[string][]gerrit.ChangeInfo{
 				"foo": {
 					{
+						Project:         "bar",
 						ID:              "1",
 						CurrentRevision: "1-1",
 						Updated:         now.Format(layout),
@@ -233,6 +230,7 @@ func TestQueryChange(t *testing.T) {
 						},
 					},
 					{
+						Project:         "bar",
 						ID:              "2",
 						CurrentRevision: "2-1",
 						Updated:         now.Add(-time.Hour).Format(layout),
@@ -244,15 +242,17 @@ func TestQueryChange(t *testing.T) {
 					},
 				},
 			},
-			revisions: []string{"1-1"},
+			revisions: map[string][]string{
+				"foo": {"1-1"},
+			},
 		},
 		{
 			name:       "multiple up-to-date changes",
-			limit:      2,
 			lastUpdate: now.Add(-time.Minute),
 			changes: map[string][]gerrit.ChangeInfo{
 				"foo": {
 					{
+						Project:         "bar",
 						ID:              "1",
 						CurrentRevision: "1-1",
 						Updated:         now.Format(layout),
@@ -263,6 +263,7 @@ func TestQueryChange(t *testing.T) {
 						},
 					},
 					{
+						Project:         "bar",
 						ID:              "2",
 						CurrentRevision: "2-1",
 						Updated:         now.Format(layout),
@@ -272,7 +273,10 @@ func TestQueryChange(t *testing.T) {
 							},
 						},
 					},
+				},
+				"baz": {
 					{
+						Project:         "boo",
 						ID:              "3",
 						CurrentRevision: "3-2",
 						Updated:         now.Format(layout),
@@ -286,6 +290,7 @@ func TestQueryChange(t *testing.T) {
 						},
 					},
 					{
+						Project:         "evil",
 						ID:              "4",
 						CurrentRevision: "4-1",
 						Updated:         now.Add(-time.Hour).Format(layout),
@@ -297,147 +302,48 @@ func TestQueryChange(t *testing.T) {
 					},
 				},
 			},
-			revisions: []string{"1-1", "2-1", "3-2"},
+			revisions: map[string][]string{
+				"foo": {"1-1", "2-1"},
+				"baz": {"3-2"},
+			},
 		},
 	}
 
 	for _, tc := range testcases {
-		fgc := &fgc{
-			changes: tc.changes,
-		}
-
-		fca := &fca{
-			c: &config.Config{
-				ProwConfig: config.ProwConfig{
-					Gerrit: config.Gerrit{
-						RateLimit: tc.limit,
+		client := &Client{
+			handlers: map[string]*gerritInstanceHandler{
+				"foo": {
+					instance: "foo",
+					projects: []string{"bar"},
+					change: &fgc{
+						changes:  tc.changes,
+						instance: "foo",
+					},
+				},
+				"baz": {
+					instance: "baz",
+					projects: []string{"boo"},
+					change: &fgc{
+						changes:  tc.changes,
+						instance: "baz",
 					},
 				},
 			},
 		}
 
-		c := &Controller{
-			ca:         fca,
-			gc:         fgc,
-			projects:   []string{"foo", "bar"},
-			lastUpdate: tc.lastUpdate,
-		}
+		changes := client.QueryChanges(tc.lastUpdate, 5)
 
-		changes := c.QueryChanges()
-
-		revisions := []string{}
-		for _, change := range changes {
-			revisions = append(revisions, change.CurrentRevision)
+		revisions := map[string][]string{}
+		for instance, changes := range changes {
+			revisions[instance] = []string{}
+			for _, change := range changes {
+				revisions[instance] = append(revisions[instance], change.CurrentRevision)
+			}
+			sort.Strings(revisions[instance])
 		}
-		sort.Strings(revisions)
 
 		if !reflect.DeepEqual(revisions, tc.revisions) {
 			t.Errorf("tc %s - wrong revisions: got %#v, expect %#v", tc.name, revisions, tc.revisions)
-		}
-	}
-}
-
-func TestProcessChange(t *testing.T) {
-	var testcases = []struct {
-		name        string
-		change      gerrit.ChangeInfo
-		numPJ       int
-		pjRef       string
-		shouldError bool
-	}{
-		{
-			name: "no revisions",
-			change: gerrit.ChangeInfo{
-				CurrentRevision: "1",
-				Project:         "test-infra",
-			},
-			shouldError: true,
-		},
-		{
-			name: "wrong project",
-			change: gerrit.ChangeInfo{
-				CurrentRevision: "1",
-				Project:         "woof",
-				Revisions: map[string]gerrit.RevisionInfo{
-					"1": {},
-				},
-			},
-		},
-		{
-			name: "normal",
-			change: gerrit.ChangeInfo{
-				CurrentRevision: "1",
-				Project:         "test-infra",
-				Revisions: map[string]gerrit.RevisionInfo{
-					"1": {
-						Ref: "refs/changes/00/1/1",
-					},
-				},
-			},
-			numPJ: 1,
-			pjRef: "refs/changes/00/1/1",
-		},
-		{
-			name: "multiple revisions",
-			change: gerrit.ChangeInfo{
-				CurrentRevision: "2",
-				Project:         "test-infra",
-				Revisions: map[string]gerrit.RevisionInfo{
-					"1": {
-						Ref: "refs/changes/00/2/1",
-					},
-					"2": {
-						Ref: "refs/changes/00/2/2",
-					},
-				},
-			},
-			numPJ: 1,
-			pjRef: "refs/changes/00/2/2",
-		},
-	}
-
-	for _, tc := range testcases {
-		fca := &fca{
-			c: &config.Config{
-				JobConfig: config.JobConfig{
-					Presubmits: map[string][]config.Presubmit{
-						"gerrit/test-infra": {
-							{
-								Name: "test-foo",
-							},
-						},
-					},
-				},
-			},
-		}
-
-		fkc := &fkc{}
-		fgc := &fgc{}
-
-		c := &Controller{
-			ca:       fca,
-			kc:       fkc,
-			gc:       fgc,
-			instance: "gerrit",
-		}
-
-		err := c.ProcessChange(tc.change)
-		if err != nil && !tc.shouldError {
-			t.Errorf("tc %s, expect no error, but got %v", tc.name, err)
-			continue
-		} else if err == nil && tc.shouldError {
-			t.Errorf("tc %s, expect error, but got none", tc.name)
-			continue
-		}
-
-		if len(fkc.prowjobs) != tc.numPJ {
-			t.Errorf("tc %s - should make %d prowjob, got %d", tc.name, tc.numPJ, len(fkc.prowjobs))
-		}
-
-		if len(fkc.prowjobs) > 0 {
-			if fkc.prowjobs[0].Spec.Refs.Pulls[0].Ref != tc.pjRef {
-				t.Errorf("tc %s - ref should be %s, got %s", tc.name, tc.pjRef, fkc.prowjobs[0].Spec.Refs.Pulls[0].Ref)
-			}
 		}
 	}
 }
